+++ web/packages/operations/package.json
@@ -19,16 +19,16 @@
     "historyV2": "npx ts-node src/global_transfer_history_v2.ts",
     "buildTokenRegistry": "npx ts-node src/build_asset_registry.ts",
     "generateReferencePreimages": "npx ts-node src/generate_reference_preimages.ts",
-    "transferWethFromPolkadotToKusama": "npx ts-node src/transfer_for_kusama_entry.ts WEthPToK 0 WETH 200000000000000",
-    "transferEthFromPolkadotToKusama": "npx ts-node src/transfer_for_kusama_entry.ts EthPToK 0 ETH 200000000000000",
-    "transferDotFromPolkadotToKusama": "npx ts-node src/transfer_for_kusama_entry.ts DotPToK 0 DOT 200000000000000",
-    "transferKsmFromPolkadotToKusama": "npx ts-node src/transfer_for_kusama_entry.ts KsmPToK 0 KSM 200000000000000",
-    "transferWethFromKusamaToPolkadot": "npx ts-node src/transfer_for_kusama_entry.ts WEthKToP 1 WETH 200000000000000",
-    "transferEthFromKusamaToPolkadot": "npx ts-node src/transfer_for_kusama_entry.ts EthKToP 1 ETH 200000000000000",
-    "transferDotFromKusamaToPolkadot": "npx ts-node src/transfer_for_kusama_entry.ts DotKToP 1 DOT 200000000000000",
-    "transferKsmFromKusamaToPolkadot": "npx ts-node src/transfer_for_kusama_entry.ts KsmKToP 1 KSM 200000000000000",
-    "transferWudFromPolkadotToKusama": "npx ts-node src/transfer_for_kusama_entry.ts WudPToK 0 WUD 200000000000000",
-    "transferWudFromKusamaToPolkadot": "npx ts-node src/transfer_for_kusama_entry.ts WudKToP 1 WUD 200000000000000",
+    "transferWethFromPolkadotToKusama": "npx ts-node src/transfer_polkadot_kusama_entry.ts WEthPToK 0 WETH 200000000000000",
+    "transferEthFromPolkadotToKusama": "npx ts-node src/transfer_polkadot_kusama_entry.ts EthPToK 0 ETH 200000000000000",
+    "transferDotFromPolkadotToKusama": "npx ts-node src/transfer_polkadot_kusama_entry.ts DotPToK 0 DOT 200000000000000",
+    "transferKsmFromPolkadotToKusama": "npx ts-node src/transfer_polkadot_kusama_entry.ts KsmPToK 0 KSM 200000000000000",
+    "transferWethFromKusamaToPolkadot": "npx ts-node src/transfer_polkadot_kusama_entry.ts WEthKToP 1 WETH 200000000000000",
+    "transferEthFromKusamaToPolkadot": "npx ts-node src/transfer_polkadot_kusama_entry.ts EthKToP 1 ETH 200000000000000",
+    "transferDotFromKusamaToPolkadot": "npx ts-node src/transfer_polkadot_kusama_entry.ts DotKToP 1 DOT 200000000000000",
+    "transferKsmFromKusamaToPolkadot": "npx ts-node src/transfer_polkadot_kusama_entry.ts KsmKToP 1 KSM 200000000000000",
+    "transferWudFromPolkadotToKusama": "npx ts-node src/transfer_polkadot_kusama_entry.ts WudPToK 0 WUD 200000000000000",
+    "transferWudFromKusamaToPolkadot": "npx ts-node src/transfer_polkadot_kusama_entry.ts WudKToP 1 WUD 200000000000000",
     "transferEtherToAH": "npx ts-node src/transfer_from_e2p.ts 1000 Eth 200000000000000",
     "transferEtherFromAH": "npx ts-node src/transfer_from_p2e.ts 1000 Eth 10000000000",
     "transferEtherFromHydration": "npx ts-node src/transfer_from_p2e.ts 2034 Eth 10000000000",
